requirements.txt
requests~=2.31.0
moralis==0.1.37
openpyxl==3.1.2
pandas==2.0.3
python-dotenv~=1.0.0